auth/requirements.txt
Flask>=1.0.0
google-cloud-firestore>=1.1
google-cloud-storage>=1.15
